Add service to track and manage related entities
Generic helper class that allows tracking of 2 related entities based on
associated entity objects and a unique identifier passed to the main helper
method. The method checks whether the related entities already exist in the
database and attaches and marks them as unchaged if they exist, or
adds them to the related context if they don't already exist in the db.